requirements.txt
Django==1.11.6
django-webpack-loader==0.5.0
djangorestframework==3.7.7
pytz==2017.2
